@@ -1,29 +1,29 @@
 {
   "name": "vue-drag-gantt-chart",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.0.0",
   "scripts": {
     "dev": "vite --host",
     "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^0.26.0",
-    "dayjs": "^1.10.7",
-    "element-ui": "^2.15.6",
+    "axios": "^1.3.5",
+    "dayjs": "^1.11.7",
+    "element-ui": "^2.15.13",
     "iscroll": "^5.2.0",
     "lodash": "^4.17.21",
     "mockjs": "^1.1.0",
     "resize-observer-polyfill": "^1.5.1",
     "v-contextmenu": "^2.9.0",
-    "vite-plugin-components": "^0.13.3",
-    "vue": "2.6.14",
+    "vue": "2.7.14",
     "vuex": "^3.6.2"
   },
   "devDependencies": {
-    "sass": "^1.49.8",
-    "vite": "^2.8.0",
-    "vite-plugin-vue2": "1.9.3",
-    "vue-template-compiler": "^2.6.14"
+    "@vitejs/plugin-vue2": "^2.2.0",
+    "sass": "^1.58.3",
+    "vite": "^4.2.1",
+    "unplugin-vue-components": "^0.24.1",
+    "vue-template-compiler": "^2.7.14"
   }
 }
